fix: filter model page and size nullable properties set to long type

+++ src/TrendyolSharp/Models/Marketplace/Filter/FilterGetOtherFinancials.cs
@@ -8,7 +8,7 @@ namespace TrendyolSharp.Models.Marketplace.Filter
     public OtherFinancialTransactionType TransactionType { get; set; }
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
-    public int? Page { get; set; }
-    public int? Size { get; set; }
+    public long? Page { get; set; }
+    public long? Size { get; set; }
   }
 }
