+++ src/TrendyolSharp/Models/Marketplace/Filter/FilterGetQuestions.cs
@@ -3,8 +3,8 @@
   public sealed class FilterGetQuestions
   {
     public long? Barcode { get; set; }
-    public int? Page { get; set; }
-    public int? Size { get; set; }
+    public long? Page { get; set; }
+    public long? Size { get; set; }
     public long? SupplierId { get; set; }
     public long? StartDate { get; set; }
     public long? EndDate { get; set; }
