+++ src/TrendyolSharp/Models/Marketplace/Filter/FilterGetShipmentPackages.cs
@@ -6,8 +6,8 @@ namespace TrendyolSharp.Models.Marketplace.Filter
   {
     public long? StartDate { get; set; }
     public long? EndDate { get; set; }
-    public int? Page { get; set; }
-    public int? Size { get; set; }
+    public long? Page { get; set; }
+    public long? Size { get; set; }
     public long? SupplierId { get; set; }
     public long? OrderNumber { get; set; }
     public string Status { get; set; }
