+++ src/TrendyolSharp/Models/Marketplace/Filter/FilterProducts.cs
@@ -6,9 +6,9 @@
     public string Barcode { get; set; }
     public long? StartDate { get; set; }
     public long? EndDate { get; set; }
-    public int? Page { get; set; }
+    public long? Page { get; set; }
     public string DateQueryType { get; set; }
-    public int? Size { get; set; }
+    public long? Size { get; set; }
     public long? SupplierId { get; set; }
     public string StockCode { get; set; }
     public bool? Archived { get; set; }
